@@ -9,43 +9,43 @@
         }
     },
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.1",
         "ext-bcmath": "*",
         "ext-calendar": "*",
         "ext-json": "*",
         "ext-pdo": "*",
-        "doctrine/annotations": "^1.13",
-        "doctrine/doctrine-bundle": "^2.4",
-        "doctrine/doctrine-fixtures-bundle": "^3.4",
-        "doctrine/doctrine-migrations-bundle": "^3.1",
-        "doctrine/event-manager": "~1.0",
-        "doctrine/orm": "^2.9",
-        "doctrine/persistence": "^1.3|^2",
+        "doctrine/annotations": "*",
+        "doctrine/dbal": "^3",
+        "doctrine/doctrine-bundle": "^2.12",
+        "doctrine/doctrine-migrations-bundle": "^3.3",
+        "doctrine/orm": "^2.19",
         "friendsofphp/proxy-manager-lts": "^1.0.2",
-        "hwi/oauth-bundle": "^1.4",
-        "jms/serializer-bundle": "^2.4",
+        "hwi/oauth-bundle": "^2.2.0",
+        "jms/serializer-bundle": "^5.4.0",
+        "lcobucci/jwt": "^4.1",
         "paypal/paypal-checkout-sdk": "^1.0.0",
         "php-http/guzzle7-adapter": "^1.0",
         "php-http/httplug-bundle": "^1.22",
-        "psr/cache": "^1.0|^2.0",
+        "psr/cache": "^3.0.0",
         "psr/container": "^1.0",
         "psr/link": "^1.0",
         "psr/log": "^1|^2",
         "qipsius/tcpdf-bundle": "^2.0",
-        "sensio/framework-extra-bundle": "^6.0",
+        "sensio/framework-extra-bundle": "^6.2",
         "stof/doctrine-extensions-bundle": "^1.6",
-        "symfony/asset": "^4.4",
-        "symfony/contracts": "^1.1.8",
-        "symfony/dotenv": "^4.4",
-        "symfony/expression-language": "^4.4",
+        "symfony/asset": "^6.4",
+        "symfony/console": "^6.4",
+        "symfony/contracts": "^3.5.0",
+        "symfony/dotenv": "^6.4",
+        "symfony/expression-language": "^6.4",
         "symfony/flex": "^1.18",
-        "symfony/form": "^4.4",
-        "symfony/http-client": "^4.4",
-        "symfony/http-foundation": "^4.4",
-        "symfony/intl": "^4.4",
+        "symfony/form": "^6.4",
+        "symfony/http-client": "^6.4",
+        "symfony/http-foundation": "^6.4",
+        "symfony/intl": "^6.4",
         "symfony/mailer": "^5.3",
         "symfony/mercure-bundle": "^0.3.4",
-        "symfony/mime": "^5.0",
+        "symfony/mime": "^6.4.7",
         "symfony/monolog-bundle": "^3.8",
         "symfony/polyfill-ctype": "~1.8",
         "symfony/polyfill-intl-grapheme": "^1.23",
@@ -56,19 +56,18 @@
         "symfony/polyfill-php73": "^1.11",
         "symfony/polyfill-php80": "^1.16",
         "symfony/polyfill-php81": "^1.22",
-        "symfony/property-access": "^4.4",
-        "symfony/property-info": "^4.4",
-        "symfony/proxy-manager-bridge": "^4.4",
-        "symfony/security-csrf": "^4.4",
-        "symfony/swiftmailer-bundle": "^3.5",
-        "symfony/templating": "^4.4",
-        "symfony/translation": "^4.4",
-        "symfony/twig-bridge": "^4.4",
-        "symfony/twig-bundle": "^4.4",
-        "symfony/validator": "^4.4",
-        "symfony/yaml": "^4.4",
-        "twig/twig": "^2.12",
-        "lcobucci/jwt": "^4.1"
+        "symfony/property-access": "^6.4",
+        "symfony/property-info": "^6.4",
+        "symfony/proxy-manager-bridge": "^6.4",
+        "symfony/routing": "^6.4",
+        "symfony/security-csrf": "^6.4",
+        "symfony/translation": "^6.4",
+        "symfony/twig-bridge": "^6.4",
+        "symfony/twig-bundle": "^6.4",
+        "symfony/validator": "^6.4",
+        "symfony/var-exporter": "^6.2",
+        "symfony/yaml": "^6.4",
+        "twig/twig": "^v3.9.3"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.3",
@@ -78,16 +77,16 @@
         "phpspec/prophecy-phpunit": "^2.0",
         "phpunit/phpunit": "^9.5",
         "psalm/plugin-symfony": "^3.1",
+        "rector/rector": "^1.0",
         "sebastian/phpcpd": "^6.0",
         "squizlabs/php_codesniffer": "2.8.1",
         "symfony/apache-pack": "^1.0",
-        "symfony/browser-kit": "^4.4",
-        "symfony/css-selector": "^4.4",
-        "symfony/debug-bundle": "^5.4",
-        "symfony/stopwatch": "^4.4",
-        "symfony/var-dumper": "^4.4",
-        "symfony/var-exporter": "^4.4",
-        "symfony/web-profiler-bundle": "^4.4",
+        "symfony/browser-kit": "^6.4",
+        "symfony/css-selector": "^6.4",
+        "symfony/debug-bundle": "^6.4",
+        "symfony/stopwatch": "^6.4",
+        "symfony/var-dumper": "^6.4",
+        "symfony/web-profiler-bundle": "^6.4",
         "vimeo/psalm": "^4.9.2",
         "weirdan/doctrine-psalm-plugin": "^2.3"
     },
