@@ -5,16 +5,16 @@ declare(strict_types=1);
 namespace App\Mealz\MealBundle\Command;
 
 use App\Mealz\MealBundle\Event\KeepAliveConnectionEvent;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
+#[AsCommand(name: 'meals:keep-alive-connection')]
 class KeepAliveConnectionCommand extends Command
 {
-    private $eventDispatcher;
-
-    protected static $defaultName = 'meals:keep-alive-connection';
+    private EventDispatcherInterface $eventDispatcher;
 
     public function __construct(EventDispatcherInterface $eventDispatcher)
     {
