@@ -70,7 +70,7 @@ class CostSheetControllerTest extends AbstractControllerTestCase
         $this->assertEquals(0.00, $balance);
 
         $this->client->request('POST', '/api/costs/settlement/' . $profile->getUsername());
-        $this->assertEquals(500, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR, $this->client->getResponse()->getStatusCode());
 
         $this->assertNull($profile->getSettlementHash(), 'Settlement was set');
     }
@@ -145,7 +145,7 @@ class CostSheetControllerTest extends AbstractControllerTestCase
 
         // Trigger action
         $this->client->request('POST', '/api/costs/hideuser/' . $profile->getUsername());
-        $this->assertEquals(500, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR, $this->client->getResponse()->getStatusCode());
 
         // Check after action
         $profile = $this->getUserProfile(parent::USER_STANDARD);
