@@ -6,12 +6,12 @@ use Symfony\Component\Asset\VersionStrategy\VersionStrategyInterface;
 
 class HashVersionStrategy implements VersionStrategyInterface
 {
-    public function getVersion($path)
+    public function getVersion($path): string
     {
         return substr(md5_file($path), 0, 7);
     }
 
-    public function applyVersion($path)
+    public function applyVersion($path): string
     {
         return true === file_exists($path) ? sprintf('%s?v=%s', $path, $this->getVersion($path)) : '';
     }
