@@ -21,6 +21,7 @@ use App\Mealz\UserBundle\Entity\Profile;
 use DateTime;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Response;
 
 class ApiController extends BaseController
 {
@@ -55,7 +56,7 @@ class ApiController extends BaseController
         return new JsonResponse([
             'paypalId' => $this->getParameter('app.paypal.client_id'),
             'mercureUrl' => $this->getParameter('app.pubsub.subscribe_url'),
-        ], 200);
+        ], Response::HTTP_OK);
     }
 
     /**
@@ -65,7 +66,7 @@ class ApiController extends BaseController
     {
         $profile = $this->getProfile();
         if (null === $profile) {
-            return new JsonResponse(null, 403);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $weeks = $this->weekSrv->getNextTwoWeeks();
@@ -148,7 +149,7 @@ class ApiController extends BaseController
             $result[$today->format('Y-m-d')] = $dishes;
         }
 
-        return new JsonResponse($result, 200);
+        return new JsonResponse($result, Response::HTTP_OK);
     }
 
     public function list(): JSONResponse
@@ -156,7 +157,7 @@ class ApiController extends BaseController
         $day = $this->apiSrv->getDayByDate(new DateTime('today'));
 
         if (null === $day) {
-            return new JsonResponse(null, 400);
+            return new JsonResponse(null, Response::HTTP_BAD_REQUEST);
         }
 
         $list['data'] = $this->participationSrv->getParticipationListBySlots($day);
@@ -173,7 +174,7 @@ class ApiController extends BaseController
 
         $list['day'] = $day->getDateTime();
 
-        return new JsonResponse($list, 200);
+        return new JsonResponse($list, Response::HTTP_OK);
     }
 
     public function listByDate(DateTime $date): JSONResponse
@@ -244,7 +245,7 @@ class ApiController extends BaseController
     {
         $profile = $this->getProfile();
         if (null === $profile) {
-            return new JsonResponse(null, 403);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $dateFrom = new DateTime('-28 days 00:00:00');
@@ -270,7 +271,7 @@ class ApiController extends BaseController
         $profile = $this->getProfile();
         $participant = $this->participationSrv->getParticipationByMealAndUser($meal, $profile);
 
-        return new JsonResponse($this->getMealState($meal, $profile, $participant), 200);
+        return new JsonResponse($this->getMealState($meal, $profile, $participant), Response::HTTP_OK);
     }
 
     private function convertMealForDashboard(Meal $meal, float $participationCount, ?Profile $profile): array
@@ -379,7 +380,7 @@ class ApiController extends BaseController
     {
         $guestInvitation = $this->guestPartiSrv->getGuestInvitationById($guestInvitationId);
         if (null === $guestInvitation) {
-            return new JsonResponse(null, 404);
+            return new JsonResponse(null, Response::HTTP_NOT_FOUND);
         }
 
         $day = $guestInvitation->getDay();
@@ -414,7 +415,7 @@ class ApiController extends BaseController
             }
         }
 
-        return new JsonResponse($guestData, 200);
+        return new JsonResponse($guestData, Response::HTTP_OK);
     }
 
     private function getDishData(Meal $meal): array
