@@ -44,7 +44,7 @@ class MealAdminControllerTest extends AbstractControllerTestCase
 
         // Request
         $this->client->request('GET', '/api/weeks');
-        $this->assertEquals(200, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_OK, $this->client->getResponse()->getStatusCode());
 
         // Get data for assertions from response
         $responseData = json_decode($this->client->getResponse()->getContent(), true);
@@ -76,7 +76,7 @@ class MealAdminControllerTest extends AbstractControllerTestCase
         $year = $date->format('Y');
         $week = $date->format('W');
         $this->createFutureEmptyWeek($date);
-        $this->assertEquals(200, $this->client->getResponse()->getStatusCode(), 'Year: ' . $year . ', week: ' . $week . ', Status: ' . $this->client->getResponse()->getContent());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_OK, $this->client->getResponse()->getStatusCode(), 'Year: ' . $year . ', week: ' . $week . ', Status: ' . $this->client->getResponse()->getContent());
 
         // Get data for assertions with new request response
         $weekRepository = $this->getDoctrine()->getRepository(Week::class);
@@ -91,7 +91,7 @@ class MealAdminControllerTest extends AbstractControllerTestCase
 
         // Trying to create the same week twice should fail
         $this->createFutureEmptyWeek($date);
-        $this->assertEquals(500, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR, $this->client->getResponse()->getStatusCode());
         $response = json_decode($this->client->getResponse()->getContent(), true);
         $this->assertEquals('102: week already exists', $response['message']);
     }
@@ -99,7 +99,7 @@ class MealAdminControllerTest extends AbstractControllerTestCase
     public function testCount(): void
     {
         $this->client->request('GET', '/api/meals/count');
-        $this->assertEquals(200, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_OK, $this->client->getResponse()->getStatusCode());
 
         $response = json_decode($this->client->getResponse()->getContent(), true);
         $this->assertNotNull($response);
@@ -118,7 +118,7 @@ class MealAdminControllerTest extends AbstractControllerTestCase
 
         // Create new week
         $this->createFutureEmptyWeek($date);
-        $this->assertEquals(200, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_OK, $this->client->getResponse()->getStatusCode());
 
         // Get data for assertions with new request response
         $weekRepository = $this->getDoctrine()->getRepository(Week::class);
@@ -189,7 +189,7 @@ class MealAdminControllerTest extends AbstractControllerTestCase
         }';
 
         $this->client->request('PUT', '/api/menu/' . $createdWeek->getId(), [], [], [], $testPutStr);
-        $this->assertEquals(200, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals(\Symfony\Component\HttpFoundation\Response::HTTP_OK, $this->client->getResponse()->getStatusCode());
 
         $createdWeek = $weekRepository->findOneBy([
             'year' => $year,
