@@ -5,16 +5,22 @@ namespace App\Mealz\UserBundle\Controller;
 use Error;
 use HWI\Bundle\OAuthBundle\Security\Core\Authentication\Token\OAuthToken;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class SecurityController extends AbstractController
 {
-    public function loginAction(Request $request): Response
+    private TokenStorageInterface $tokenStorage;
+    public function __construct(TokenStorageInterface $tokenStorage)
+    {
+        $this->tokenStorage = $tokenStorage;
+    }
+    public function login(Request $request): Response
     {
         // If Keycloak is enabled, redirect to the Meals home
-        $token = $this->get('security.token_storage')->getToken();
+        $token = $this->tokenStorage->getToken();
         if ($token instanceof OAuthToken) {
             return $this->redirectToRoute('MealzMealBundle_home');
         }
