@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace App\Mealz\UserBundle\Provider;
 
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Security\Http\Logout\LogoutSuccessHandlerInterface;
+use Symfony\Component\Security\Http\Event\LogoutEvent;
 
 /**
  * Class LogoutSuccessHandler.
  */
-class LogoutSuccessHandler implements LogoutSuccessHandlerInterface
+class LogoutSuccessHandler implements EventSubscriberInterface
 {
     private string $logoutUrl;
 
@@ -22,21 +22,36 @@ class LogoutSuccessHandler implements LogoutSuccessHandlerInterface
     }
 
     /**
-     * @return JsonResponse|RedirectResponse
+     * Redirects on successful logout.
      */
-    public function onLogoutSuccess(Request $request)
+    public function onLogout(LogoutEvent $event): void
     {
+        $request = $event->getRequest();
+
         if ($request->isXmlHttpRequest()) {
             $response = new JsonResponse([
                 'payload' => null,
                 'error' => [],
                 'redirect' => $this->logoutUrl,
             ]);
-            $response->setStatusCode(302);
+            $response->setStatusCode(\Symfony\Component\HttpFoundation\Response::HTTP_FOUND);
+            $event->setResponse($response);
 
-            return $response;
+            return;
         }
 
-        return new RedirectResponse($this->logoutUrl);
+        $response = new RedirectResponse($this->logoutUrl);
+        $event->setResponse($response);
+    }
+
+    /**
+     * @inheritdoc
+     * @return string[]
+     */
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            LogoutEvent::class => 'onLogout',
+        ];
     }
 }
